backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
yfinance==0.2.44
pandas==2.2.3
TA-Lib==0.4.32
textblob==0.18.0.post0
requests==2.32.3
tensorflow==2.17.0
matplotlib==3.9.2
numpy==1.26.4
scikit-learn==1.5.2
python-dotenv==1.0.1
sqlalchemy==2.0.35
asyncpg==0.30.0
psycopg2-binary==2.9.9
apscheduler==3.10.4